# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Middleware for managing Mikrotik Routers",
        "title": "NetEngine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Checks if the NetEngine Muscle is alive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get API Health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/isolate": {
            "post": {
                "description": "Adds or removes a user from a Firewall Address List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced"
                ],
                "summary": "Isolate User",
                "parameters": [
                    {
                        "description": "Isolation Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IsolateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/router/{id}/health": {
            "get": {
                "description": "Returns CPU, Memory, and Uptime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get Router Health",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Router ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SystemResource"
                        }
                    }
                }
            }
        },
        "/router/{id}/users": {
            "get": {
                "description": "Returns all PPPoE users from database with real-time connection status (connected, isolated, offline)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get All Users with Status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Router ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserWithStatus"
                            }
                        }
                    }
                }
            }
        },
        "/secret": {
            "post": {
                "description": "Adds a new PPPoE secret to a router",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "Create PPP Secret",
                "parameters": [
                    {
                        "description": "Secret Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateSecretRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/secret/{user}": {
            "put": {
                "description": "Updates the profile of an existing secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bridge"
                ],
                "summary": "Change Plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Plan Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sync/{id}": {
            "post": {
                "description": "Triggers an immediate sync of active users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "Force Sync Router",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Router ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateSecretRequest": {
            "type": "object",
            "required": [
                "password",
                "profile",
                "user"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "local_ip": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "remote_ip": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "api.IsolateRequest": {
            "type": "object",
            "required": [
                "action",
                "ip"
            ],
            "properties": {
                "action": {
                    "description": "add or remove",
                    "type": "string",
                    "enum": [
                        "add",
                        "remove"
                    ]
                },
                "comment": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "list": {
                    "description": "Default to \"ISOLATED\" if empty",
                    "type": "string"
                }
            }
        },
        "api.UpdatePlanRequest": {
            "type": "object",
            "required": [
                "profile"
            ],
            "properties": {
                "profile": {
                    "type": "string"
                }
            }
        },
        "models.SystemResource": {
            "type": "object",
            "properties": {
                "board_name": {
                    "type": "string"
                },
                "cpu": {
                    "description": "Load in %",
                    "type": "string"
                },
                "free_memory": {
                    "type": "integer"
                },
                "total_memory": {
                    "type": "integer"
                },
                "uptime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.UserWithStatus": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "status": {
                    "description": "\"connected\", \"isolated\", or \"offline\"",
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AppKey": {
            "type": "apiKey",
            "name": "X-App-Key",
            "in": "header"
        }
    }
}