# internal/models/pppoe.go
package models

// PPPoESecret represents a PPPoE account from MikroTik
type PPPoESecret struct {
	Name          string
	Profile       string
	RemoteAddress string
	Disabled      bool
}

// UserWithStatus represents a user with their connection status
type UserWithStatus struct {
	Username string `json:"username"`
	Status   string `json:"status"` // "connected", "isolated", or "offline"
	IP       string `json:"ip,omitempty"`
	Uptime   string `json:"uptime,omitempty"`
	Profile  string `json:"profile,omitempty"`
}